add geometry info into BlockDataContainer
